@@ -3,7 +3,7 @@ package lab2.lab2.Service;
 import java.util.List;
 import java.util.Optional;
 
-import org.springframework.beans.factory.annotation.Autowired;
+// import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import lab2.lab2.Model.Student;
@@ -11,8 +11,11 @@ import lab2.lab2.Repository.StudentRepo;
 
 @Service
 public class StudentService {
-    @Autowired
-    private final StudentRepo studentRepo = new StudentRepo();
+    private final StudentRepo studentRepo;
+
+    public StudentService(StudentRepo studentRepo) {
+        this.studentRepo = studentRepo;
+    }
 
     public Student save(Student student) {
         return studentRepo.save(student);
@@ -28,10 +31,6 @@ public class StudentService {
 
     public Student update(Long id, Student student) {
         return studentRepo.update(id, student);
-    }
-
-    public void delete(Student student) {
-        studentRepo.delete(student);
     }
 
     public void deleteById(Long id) {
